+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -751,7 +751,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6AYYBBRC1HN21DMWRKR7BWS"),
+  _id: z.string().default("01K6B3TW075N0K3T47SWPXZ6QE"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -795,7 +795,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K6AYYBBRC1HN21DMWRKR7BWS"),
+  id: z.string().default("01K6B3TW075N0K3T47SWPXZ6QE"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,