+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -172,7 +172,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6AYYBBRF3Q1VC4GMRR14MCG"),
+  _id: z.string().default("01K6B3TW06CQ8YR498JFMRH7BF"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -216,7 +216,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K6AYYBBRF3Q1VC4GMRR14MCG"),
+  id: z.string().default("01K6B3TW06CQ8YR498JFMRH7BF"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,